+++ src/tests/features/preferPromiseFnWhenKnown.test.ts
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest"
 
-import { getDTSFilesForRun, graphql, prisma } from "../testRunner"
+import { getDTSFilesForRun, graphql, prisma } from "../testRunner.js"
 
 it("uses a rn to promise when we see an async tag", () => {
 	const prismaSchema = prisma`
@@ -50,53 +50,103 @@ export const Game = {
 	const { vfsMap } = getDTSFilesForRun({ sdl, gamesService, prismaSchema })
 	const dts = vfsMap.get("/types/games.d.ts")!
 	expect(dts.trim()).toMatchInlineSnapshot(`
-		"import type { Query } from \\"./shared-schema-types\\";
-		import type { Game as RTGame } from \\"./shared-return-types\\";
-		import type { Game as PGame } from \\"@prisma/client\\";
+		"import type { Game as PGame } from \\"@prisma/client\\";
 		import type { GraphQLResolveInfo } from \\"graphql\\";
-		import type { RedwoodGraphQLContext } from \\"@redwoodjs/graphql-server/dist/functions/types\\";
+
+		import type { RedwoodGraphQLContext } from \\"@redwoodjs/graphql-server/dist/types\\";
+
+		import type { Game as RTGame } from \\"./shared-return-types\\";
+		import type { Query } from \\"./shared-schema-types\\";
 
 		/** SDL: gameSync: Game */
 		export interface GameSyncResolver {
-		  (args?: object, obj?: { root: Query, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }): RTGame | null | Promise<RTGame | null> | (() => Promise<RTGame | null>);
+		  (
+		    args?: object,
+		    obj?: {
+		      root: Query;
+		      context: RedwoodGraphQLContext;
+		      info: GraphQLResolveInfo;
+		    }
+		  ): RTGame | null | Promise<RTGame | null> | (() => Promise<RTGame | null>);
 		}
 
 		/** SDL: gameAsync: Game */
 		export interface GameAsyncResolver {
-		  (args?: object, obj?: { root: Query, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }): Promise<RTGame | null>;
+		  (
+		    args?: object,
+		    obj?: {
+		      root: Query;
+		      context: RedwoodGraphQLContext;
+		      info: GraphQLResolveInfo;
+		    }
+		  ): Promise<RTGame | null>;
 		}
 
 		/** SDL: gameAsync1Arg: Game */
 		export interface GameAsync1ArgResolver {
-		  (args: object, obj?: { root: Query, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }): RTGame | null | Promise<RTGame | null> | (() => Promise<RTGame | null>);
+		  (
+		    args: object,
+		    obj?: {
+		      root: Query;
+		      context: RedwoodGraphQLContext;
+		      info: GraphQLResolveInfo;
+		    }
+		  ): RTGame | null | Promise<RTGame | null> | (() => Promise<RTGame | null>);
 		}
 
 		/** SDL: gameAsync2Arg: Game */
 		export interface GameAsync2ArgResolver {
-		  (args: object, obj: { root: Query, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }): RTGame | null | Promise<RTGame | null> | (() => Promise<RTGame | null>);
+		  (
+		    args: object,
+		    obj: {
+		      root: Query;
+		      context: RedwoodGraphQLContext;
+		      info: GraphQLResolveInfo;
+		    }
+		  ): RTGame | null | Promise<RTGame | null> | (() => Promise<RTGame | null>);
 		}
 
 		/** SDL: gameObj: Game */
 		export interface GameObjResolver {
-		  (args?: object, obj?: { root: Query, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }): RTGame | null;
+		  (
+		    args?: object,
+		    obj?: {
+		      root: Query;
+		      context: RedwoodGraphQLContext;
+		      info: GraphQLResolveInfo;
+		    }
+		  ): RTGame | null;
 		}
 
 		export interface GameTypeResolvers<Extended> {
-
 		  /** SDL: summary: String! */
 		  summary: string;
 
 		  /** SDL: summarySync: String! */
-		  summarySync: (args?: undefined, obj?: { root: GameAsParent<Extended>, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }) => string | Promise<string> | (() => Promise<string>);
+		  summarySync: (
+		    args?: undefined,
+		    obj?: {
+		      root: GameAsParent<Extended>;
+		      context: RedwoodGraphQLContext;
+		      info: GraphQLResolveInfo;
+		    }
+		  ) => string | Promise<string> | (() => Promise<string>);
 
 		  /** SDL: summaryAsync: String! */
-		  summaryAsync: (args?: undefined, obj?: { root: GameAsParent<Extended>, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }) => Promise<string>;
+		  summaryAsync: (
+		    args?: undefined,
+		    obj?: {
+		      root: GameAsParent<Extended>;
+		      context: RedwoodGraphQLContext;
+		      info: GraphQLResolveInfo;
+		    }
+		  ) => Promise<string>;
 		}
 
 		type GameAsParent<Extended> = PGame & {
-		  summary: () => string,
-		  summarySync: () => string | Promise<string> | (() => Promise<string>),
-		  summaryAsync: () => Promise<string>
+		  summary: () => string;
+		  summarySync: () => string | Promise<string> | (() => Promise<string>);
+		  summaryAsync: () => Promise<string>;
 		} & Extended;"
 	`)
 })