+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 # Changelog
+
+### 0.0.10
+
+- Service files do not trigger eslint warnings in (my) app OOTB
+- Adds support for running prettier in the generated files.
+- Fixes paths for the graphql types which come from `@redwoodjs/graphql-server`
+- Does not create empty service .d.ts files